facebook-groups-post-bot/requirements.txt
selenium==4.12.0
python-dotenv==1.0.0
undetected-chromedriver~=3.5.5